move useRefetchPriceCallback to new folder
- moved "util"esque files to own utils.ts
- make getBestPriceQuote to operator file
